Rename MQTT Agent file
