+++ source/simple_sub_pub_demo.c
@@ -58,7 +58,7 @@
 #include "core_mqtt.h"
 
 /* MQTT agent include. */
-#include "freertos_mqtt_agent.h"
+#include "mqtt_agent.h"
 
 /* Subscription manager header include. */
 #include "subscription_manager.h"
